chore: 프로젝트 폴더 구조 개선 (#78)
- vehicle 폴더명을 log로 변경
